+++ src/com/jpexs/decompiler/flash/easygui/EasySwfPanel.java
@@ -532,7 +532,7 @@ public class EasySwfPanel extends JPanel {
                 if (obj instanceof Tag) {
                     Tag t = (Tag) obj;
                     libraryPreviewPanel.setTimelined(TimelinedMaker.makeTimelined(t), t.getSwf(),
-                            -1, false, true, true, true, true, false, true);
+                            -1, false, true, true, true, true, false, true, true);
                     libraryPreviewPanel.zoomFit();
                 } else {
                     libraryPreviewPanel.clearAll();
@@ -579,7 +579,7 @@ public class EasySwfPanel extends JPanel {
             libraryTreeTable.setSwf(swf);
             libraryPreviewPanel.clearAll();
             if (updateStage) {
-                stagePanel.setTimelined(timelined, swf, 0, true, true, true, true, true, false, true);
+                stagePanel.setTimelined(timelined, swf, 0, true, true, true, true, true, false, true, true);
                 stagePanel.pause();
                 stagePanel.gotoFrame(0);
             }